requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
arcade-ai==2.1.1
arcade-core==2.2.0
arcade-tdk==2.1.0
arcade_math==2.0.2
arcadepy==1.5.0
asgiref==3.9.1
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
bs4==0.0.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
Flask==3.1.1
frozenlist==1.7.0
fsspec==2025.7.0
googleapis-common-protos==1.70.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.6
jira==3.5.2
jiter==0.10.0
joblib==1.5.1
llama-index==0.10.0
llama-index-agent-openai==0.1.0
llama-index-core==0.10.0
llama-index-embeddings-openai==0.1.0
llama-index-legacy==0.9.48.post4
llama-index-llms-openai==0.1.0
llama-index-multi-modal-llms-openai==0.1.0
llama-index-program-openai==0.1.0
llama-index-question-gen-openai==0.1.0
llama-index-readers-file==0.1.0
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
multidict==6.6.3
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
numpy==2.3.2
oauthlib==3.3.1
openai==1.82.1
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-http==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-instrumentation-asgi==0.49b2
opentelemetry-instrumentation-fastapi==0.49b2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
opentelemetry-util-http==0.49b2
packaging==25.0
pandas==2.3.1
pillow==11.3.0
propcache==0.3.2
protobuf==5.29.5
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
PyMuPDF==1.26.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
six==1.17.0
slack_sdk==3.27.1
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
tenacity==8.5.0
tiktoken==0.9.0
toml==0.10.2
tqdm==4.67.1
typer==0.10.0
types-python-dateutil==2.9.0.20241003
types-pytz==2024.2.0.20241003
types-toml==0.10.8.20240310
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
win32_setctime==1.2.0
wrapt==1.17.2
yarl==1.20.1
zipp==3.23.0
